ai_service/requirements.txt
mcp 
httpx 
chromadb 
openai 
streamlit 
python-dotenv